SH-8 Init core modules in app module
