+++ src/core/index.ts
@@ -1,0 +1,2 @@
+export * from './svg-icons/svg-icons.module';
+export * from './translation-config/translation-config.module';
