+++ src/core/svg-icons/components/index.ts
@@ -1,0 +1,1 @@
+export * from './svg-icon/svg-icon.component';
