+++ src/core/svg-icons/services/index.ts
@@ -1,0 +1,2 @@
+export * from './sprite-loader/sprite-loader.service';
+export * from './svg/svg.service';
