+++ src/core/translation-config/services/index.ts
@@ -1,0 +1,1 @@
+export * from './translation-loader/translation-loader.service';
