+++ src/core/translation-config/translation-config.module.ts
@@ -1,0 +1,34 @@
+import { APP_INITIALIZER, NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { TranslationLoaderService } from './services';
+
+export function HttpLoaderFactory(http: HttpClient): TranslateHttpLoader {
+  return new TranslateHttpLoader(http, './assets/i18n/', '.json');
+}
+
+@NgModule({
+  imports: [
+    CommonModule,
+    HttpClientModule,
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient],
+      },
+    }),
+  ],
+  providers: [
+    TranslationLoaderService,
+    {
+      provide: APP_INITIALIZER,
+      useFactory: (translationLoader: TranslationLoaderService) => () => translationLoader.loadTranslation(),
+      deps: [TranslationLoaderService],
+      multi: true,
+    },
+  ]
+})
+export class TranslationConfigModule {}
