+++ src/shared/components/index.ts
@@ -1,6 +1,5 @@
 export * from './header/header.component';
 export * from './navbar/navbar.component';
-export * from './svg-icon/svg-icon.component';
 export * from './searchbar/searchbar.component';
 export * from './loader/loader.component';
 export * from './cart-icon/cart-icon.component';
