+++ src/shared/services/index.ts
@@ -1,6 +1,3 @@
-export * from './sprite-loader.service';
-export * from './svg.service';
-export * from './translator-loader.service';
 export * from './search-facade.service';
 export * from './search-facade.service';
 export * from './brand';
